requirements.txt
apt-xapian-index==0.49
argcomplete==1.8.1
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.1
blinker==1.4
bson==0.5.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.1
entrypoints==0.3
fastapi==0.60.1
google==3.0.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gpg===1.13.1-unknown
grpcio==1.31.0
h11==0.9.0
httplib2==0.18.1
httptools==0.1.1
idna==2.10
ImageHash==4.1.0
isort==4.3.21
joblib==0.16.0
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
meld==3.20.2
mock==3.0.5
netifaces==0.10.4
numpy==1.19.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
pandas==1.1.0
pbr==5.4.5
pexpect==4.6.0
Pillow==8.0.1
progressbar==2.5
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.14.1
python-apt==2.0.0+ubuntu0.20.4.2
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-gnupg==0.4.5
python-jose==3.2.0
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
QtPy==1.9.0
reportlab==3.5.34
requests==2.24.0
requests-unixsocket==0.2.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.15.0
soupsieve==2.0.1
starlette==0.13.6
systemd-python==234
threadpoolctl==2.1.0
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
Ubuntu-Make==18.9
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.10
usb-creator==0.3.7
uvicorn==0.11.6
uvloop==0.14.0
wadllib==1.3.3
websockets==8.1
wrapt==1.12.1
xkit==0.0.0
